requirements/dev.txt
-r base.txt

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
